backend/requirements.txt
fastapi
uvicorn
python-jose
bcrypt==4.0.1
passlib[bcrypt]
